prime-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt<4
PyJWT>=2.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
